edge-AI/requirements.txt
paho-mqtt
confluent-kafka
numpy
scipy
joblib
pandas
scikit-learn==1.6.1